# Conflicts:
#	split/models.py
#	templates/cost_edit.html
#	templates/cost_view.html

+++ split/models.py
@@ -4,12 +4,11 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 
 # Create your models here.
-from django.utils import timezone
 
 
 class Profile(models.Model):
     user = models.OneToOneField(User,  on_delete=models.CASCADE)
-    balance = models.DecimalField(max_digits=14, decimal_places=2, default=0.0000)
+    balance = models.FloatField(default=0.0)
 
     def __str__(self):
         return self.user.username
@@ -35,7 +34,7 @@ class Group(models.Model):
 class GroupUser(models.Model):
     group_id = models.ForeignKey("Group", on_delete=models.CASCADE)
     user_id = models.ForeignKey("Profile", on_delete=models.CASCADE)
-    balance = models.DecimalField(max_digits=14, decimal_places=2, default=0.0000)
+    balance = models.FloatField(default=0.0)
 
     def __str__(self):
         return f"{self.group_id} - {self.user_id} - {self.balance}"
@@ -43,7 +42,7 @@ class GroupUser(models.Model):
 
 class Cost(models.Model):
     title = models.CharField(max_length=100)
-    amount = models.DecimalField(max_digits=14, decimal_places=2)
+    amount = models.FloatField()
     payer_id = models.ForeignKey("Profile", on_delete=models.DO_NOTHING)
     group_id = models.ForeignKey("Group", on_delete=models.CASCADE)
     users = models.ManyToManyField("Profile", related_name="cost_users_many_to_many")
